refactor: Headers 객체분리
